requirements.txt
robotframework>=7.0
robotframework-requests>=0.9
robotframework-pabot>=2.16
robotframework-robocop>=5.0
pyyaml>=6.0
requests>=2.31
